feat: once gap check is complete, generate summary JSON file

+++ src/check_gaps.py
@@ -1,13 +1,15 @@
 import csv
+import json
 import os
 import re
+from datetime import datetime
 from io import TextIOWrapper
 from pathlib import Path
 from typing import Optional
-from constants import CONSTANTS
 
 import typer
 
+from constants import CONSTANTS
 from science_mode import Constants, Mode, ModeConfig
 
 app = typer.Typer()
@@ -45,6 +47,16 @@ def main(
     no_report: bool = typer.Option(
         False, "--no-report", help="Disable the report file"
     ),
+    report_file_suffix: str = typer.Option(
+        ".gap_report.txt",
+        "--report-suffix",
+        help="Suffix used to identify report file names and automatically appended to data file names when generating report names automatically",
+    ),
+    summarise_only: bool = typer.Option(
+        False,
+        "--summarise",
+        help="Skip the gap checking and just generate the summary files",
+    ),
 ):
     """
     Check MAG science CSV files for gaps in sequence counters and time stamps
@@ -54,135 +66,151 @@ def main(
     global exit_code
 
     if not no_report and not report_file_path.name:
-        report_file_path = Path(f"{data_file.name.replace('.csv', '')}.gap_report.txt")
+        report_file_path = Path(
+            f"{data_file.name.replace('.csv', '')}{report_file_suffix}"
+        )
+    # if the reportname has been specified and no suffix has been specified then use the report file name as the suffix
+    elif report_file_suffix == ".gap_report.txt":
+        report_file_suffix = report_file_path.name
 
-    mode = validate_check_gap_args(data_file, report_file_path, mode, force, no_report)
-
-    if not no_report:
-        report_file = open(report_file_path, "a")
-    else:
-        no_report_flag = no_report
-
-    reader = csv.DictReader(data_file)
-
-    if mode != Mode.auto:
-        mode_config = ModeConfig(mode)
-    else:
-        mode_config = ModeConfig(data_file.name)
-
-    exit_code = 0
-    line_count = 0
-    packet_start_line_count = 0
-    packet_line_count = 0
-    packet_counter = 0
-    prev_seq = -1
-    primary_vector_count = 0
-    secondary_vector_count = 0
-
-    write_line(
-        f"Checking {data_file.name} in mode {mode.value} ({mode_config.primary_rate}, {mode_config.secondary_rate}) @ {mode_config.seconds_between_packets}s"
+    mode = validate_check_gap_args(
+        data_file, report_file_path, mode, force, no_report, summarise_only
     )
 
-    for row in reader:
-        packet_line_count += 1
-        line_count += 1
-
-        sequence = get_integer(line_count, row, "sequence")
-
-        if line_count == 1 and packet_counter == 0:
-            # we have our first packet
-            packet_counter += 1
-            packet_start_line_count += 1
-
-        # if the seq count has moved we assume this is the start of a new packet
-        if line_count > 1 and sequence != prev_seq:
-            # check the previous packet has a complete set of vectors
-            verify_packet_completeness(
-                primary_vector_count,
-                secondary_vector_count,
-                mode_config,
-                prev_seq,
-                packet_start_line_count,
-                False,
-            )
-            packet_counter += 1
-            packet_start_line_count = line_count
-            primary_vector_count = 0
-            secondary_vector_count = 0
-
-        packet_line_count = verify_sequence_counter(
-            mode_config, line_count, packet_line_count, prev_seq, sequence
-        )
-
-        hasPrimary = packet_line_count <= mode_config.primary_vectors_per_packet
-        hasSeconday = packet_line_count <= mode_config.secondary_vectors_per_packet
-
-        if packet_line_count > mode_config.rows_per_packet:
-            if packet_line_count == mode_config.rows_per_packet + 1:
-                write_error(
-                    f"{CONSTANTS.TOO_MANY_ROWS}. Expected {mode_config.rows_per_packet}. line number {line_count + 1}, sequence count: {sequence}"
-                )
+    if not summarise_only:
+        if not no_report:
+            report_file = open(report_file_path, "a")
         else:
-            if hasPrimary:
-                pri_coarse = get_integer(line_count, row, "pri_coarse")
-                pri_fine = get_integer(line_count, row, "pri_fine")
+            no_report_flag = no_report
 
-                verify_timestamp(
-                    mode_config,
-                    line_count,
-                    packet_line_count,
-                    sequence,
-                    pri_coarse,
-                    pri_fine,
-                    "primary",
-                )
-                if verify_non_zero_vectors(row, line_count, sequence, "primary"):
-                    primary_vector_count += 1
-            else:
-                verify_empty_vectors(row, line_count, sequence, "primary")
+        reader = csv.DictReader(data_file)
 
-            if hasSeconday:
-                sec_coarse = get_integer(line_count, row, "sec_coarse")
-                sec_fine = get_integer(line_count, row, "sec_fine")
+        if mode != Mode.auto:
+            mode_config = ModeConfig(mode)
+        else:
+            mode_config = ModeConfig(data_file.name)
 
-                verify_timestamp(
-                    mode_config,
-                    line_count,
-                    packet_line_count,
-                    sequence,
-                    sec_coarse,
-                    sec_fine,
-                    "secondary",
-                )
-                if verify_non_zero_vectors(row, line_count, sequence, "secondary"):
-                    secondary_vector_count += 1
-            else:
-                verify_empty_vectors(row, line_count, sequence, "secondary")
+        exit_code = 0
+        line_count = 0
+        packet_start_line_count = 0
+        packet_line_count = 0
+        packet_counter = 0
+        prev_seq = -1
+        primary_vector_count = 0
+        secondary_vector_count = 0
 
-        prev_seq = sequence
-
-    # check the last packet has a complete set of vectors
-    verify_packet_completeness(
-        primary_vector_count,
-        secondary_vector_count,
-        mode_config,
-        prev_seq,
-        packet_start_line_count,
-        True,
-    )
-
-    if exit_code != 0:
         write_line(
-            f"Error - found bad science data! Checked {packet_counter} packet(s) across {line_count} rows of data."
+            f"Checking {data_file.name} in mode {mode.value} ({mode_config.primary_rate}, {mode_config.secondary_rate}) @ {mode_config.seconds_between_packets}s"
         )
-    else:
-        write_line(
-            f"Gap checker completed successfully. Checked {packet_counter} packet(s) across {line_count} rows of data."
+
+        for row in reader:
+            packet_line_count += 1
+            line_count += 1
+
+            sequence = get_integer(line_count, row, "sequence")
+
+            if line_count == 1 and packet_counter == 0:
+                # we have our first packet
+                packet_counter += 1
+                packet_start_line_count += 1
+
+            # if the seq count has moved we assume this is the start of a new packet
+            if line_count > 1 and sequence != prev_seq:
+                # check the previous packet has a complete set of vectors
+                verify_packet_completeness(
+                    primary_vector_count,
+                    secondary_vector_count,
+                    mode_config,
+                    prev_seq,
+                    packet_start_line_count,
+                    False,
+                )
+                packet_counter += 1
+                packet_start_line_count = line_count
+                primary_vector_count = 0
+                secondary_vector_count = 0
+
+            packet_line_count = verify_sequence_counter(
+                mode_config, line_count, packet_line_count, prev_seq, sequence
+            )
+
+            hasPrimary = packet_line_count <= mode_config.primary_vectors_per_packet
+            hasSeconday = packet_line_count <= mode_config.secondary_vectors_per_packet
+
+            if packet_line_count > mode_config.rows_per_packet:
+                if packet_line_count == mode_config.rows_per_packet + 1:
+                    write_error(
+                        f"{CONSTANTS.TOO_MANY_ROWS}. Expected {mode_config.rows_per_packet}. line number {line_count + 1}, sequence count: {sequence}"
+                    )
+            else:
+                if hasPrimary:
+                    pri_coarse = get_integer(line_count, row, "pri_coarse")
+                    pri_fine = get_integer(line_count, row, "pri_fine")
+
+                    verify_timestamp(
+                        mode_config,
+                        line_count,
+                        packet_line_count,
+                        sequence,
+                        pri_coarse,
+                        pri_fine,
+                        "primary",
+                    )
+                    if verify_non_zero_vectors(row, line_count, sequence, "primary"):
+                        primary_vector_count += 1
+                else:
+                    verify_empty_vectors(row, line_count, sequence, "primary")
+
+                if hasSeconday:
+                    sec_coarse = get_integer(line_count, row, "sec_coarse")
+                    sec_fine = get_integer(line_count, row, "sec_fine")
+
+                    verify_timestamp(
+                        mode_config,
+                        line_count,
+                        packet_line_count,
+                        sequence,
+                        sec_coarse,
+                        sec_fine,
+                        "secondary",
+                    )
+                    if verify_non_zero_vectors(row, line_count, sequence, "secondary"):
+                        secondary_vector_count += 1
+                else:
+                    verify_empty_vectors(row, line_count, sequence, "secondary")
+
+            prev_seq = sequence
+
+        # check the last packet has a complete set of vectors
+        verify_packet_completeness(
+            primary_vector_count,
+            secondary_vector_count,
+            mode_config,
+            prev_seq,
+            packet_start_line_count,
+            True,
         )
+
+        if exit_code != 0:
+            write_line(
+                f"Error - found bad science data! Checked {packet_counter} packet(s) across {line_count} rows of data."
+            )
+        else:
+            write_line(
+                f"Gap checker completed successfully. Checked {packet_counter} packet(s) across {line_count} rows of data."
+            )
+
+        if report_file:
+            report_file.close()
+
+        if not no_report:
+            print(f"Report saved to {report_file_path}")
+    # end of if not summarise_only
 
     if not no_report:
-        report_file.close()
-        print(f"Report saved to {report_file_path}")
+        # generate a nice sumary of all ther errors by scanning all gap reports in the folder
+        folder = Path(data_file.name).parent
+        generate_summary(folder, f"*{report_file_suffix}")
 
     if exit_code != 0:
         raise typer.Exit(code=exit_code)
@@ -199,8 +227,10 @@ def get_integer(line_count, row, field):
     return value
 
 
-def validate_check_gap_args(data_file, report_file_path, mode, force, no_report):
-    if not (no_report) and report_file_path.exists():
+def validate_check_gap_args(
+    data_file, report_file_path, mode, force, no_report, summarise_only
+):
+    if not (no_report) and report_file_path.exists() and not summarise_only:
         if force:
             os.remove(report_file_path)
         else:
@@ -387,6 +417,77 @@ def write_error(message: str):
     write_line(message)
 
 
+def generate_summary(folder: Path, report_file_glob: str):
+    """Scan the folder for report files that match report_file_suffic and generate a summary count of all the errors"""
+
+    global exit_code
+
+    # get a list of all files in folder that match the report_file_suffix
+    report_files = list(folder.glob(report_file_glob))
+    summaryJsonPath = Path(folder, "gap_check_summary.json")
+
+    if not report_files:
+        print(
+            f"No report file(s) matching {report_file_glob} found in {folder} so unable to generate a summary"
+        )
+        exit_code = 3
+        return
+
+    print(
+        f"{len(report_files)} file(s) matching {report_file_glob} found in {folder} - generating {summaryJsonPath}"
+    )
+
+    # iterate over each report file and generate a summary
+    summary = {}
+    summary["Folder"] = str(folder.absolute())
+    summary["Generated"] = str(datetime.now())
+
+    for file in report_files:
+        with open(file, "r") as f:
+            for line in f:
+                error = None
+                if line.find(CONSTANTS.VECTORS_ALL_ZERO) != -1:
+                    error = "Vectors are all zero errors"
+                elif line.find(CONSTANTS.NONE_SEQUENTIAL) != -1 or line.startswith(
+                    f"A {CONSTANTS.PACKET_INCOMPLETE}"
+                ):  # ignore last line error
+                    error = "Missing science data errors"
+                elif (
+                    re.search(CONSTANTS.EXPECTED_NUMERIC_MATCH_REGEX, line)
+                    or line.find(CONSTANTS.SEQUENCE_NUMBERS_VARY) != -1
+                    or line.find(CONSTANTS.RANGE_IS_INVALID) != -1
+                    or line.find(CONSTANTS.VECTORS_NON_EMPTY) != -1
+                    or line.find(CONSTANTS.TOO_MANY_ROWS) != -1
+                    or line.find(CONSTANTS.PACKET_TOO_BIG) != -1
+                ):
+                    error = "Corrupt science packet errors"
+
+                elif line.find(CONSTANTS.TIMESTAMP) != -1:
+                    error = "incorrect timestamp errors"
+
+                if error:
+                    if error in summary:
+                        summary[error] += 1
+                    else:
+                        summary[error] = 1
+
+            f.close()
+
+    if len(summary) == 2:
+        summary["Gap check result"] = "PASSED"
+    else:
+        summary["Gap check result"] = "FAILED"
+
+    if summaryJsonPath.exists():
+        os.remove(summaryJsonPath)
+
+    with open(summaryJsonPath, "w") as fp:
+        json.dump(summary, fp, indent=2)  # encode dict into JSON
+
+
 # only needed when this file is run as its own app
 if __name__ == "__main__":
     app()
+
+    if report_file:
+        report_file.close()
